#872 - runtime: clean up streams
